[ADD] Setting for smooth movement, respect system settings by default
