+++ pandora-client-web/src/graphics/graphicsSettings.tsx
@@ -1,18 +1,15 @@
-import { CloneDeepMutable, FormatBytes } from 'pandora-common';
-import React, { ReactElement, useCallback, useEffect, useMemo, useState } from 'react';
+import { AssertNever, CloneDeepMutable } from 'pandora-common';
+import { useMemo } from 'react';
 import { z } from 'zod';
-import { GraphicsManagerInstance, type IGraphicsLoaderStats } from '../assets/graphicsManager';
+import { GraphicsManagerInstance } from '../assets/graphicsManager';
 import { BrowserStorage } from '../browserStorage';
-import { Button } from '../components/common/button/button';
-import { Column, Row } from '../components/common/container/container';
-import { ContextHelpButton } from '../components/help/contextHelpButton';
-import { SelectSettingInput, ToggleSettingInput } from '../components/settings/helpers/settingsInputs';
+import { useMediaQuery } from '../common/useMediaQuery';
 import { useObservable } from '../observable';
-import { useAutomaticResolution } from '../services/screenResolution/screenResolution';
 
 export const GraphicsSettingsSchema = z.object({
 	// Effects
 	effectBlinking: z.boolean(),
+	smoothMovement: z.enum(['auto', 'enabled', 'disabled']),
 	// Quality
 	renderResolution: z.number().int().min(0).max(100),
 	textureResolution: z.enum(['auto', '1', '0.5', '0.25']),
@@ -20,19 +17,20 @@ export const GraphicsSettingsSchema = z.object({
 });
 export type GraphicsSettings = z.infer<typeof GraphicsSettingsSchema>;
 
-const GRAPHICS_SETTINGS_DEFAULT: GraphicsSettings = {
+export const GRAPHICS_SETTINGS_DEFAULT: Readonly<GraphicsSettings> = {
 	// Effects
 	effectBlinking: true,
+	smoothMovement: 'auto',
 	// Quality
 	renderResolution: 100,
 	textureResolution: 'auto',
 	alphamaskEngine: 'disabled',
-};
+} as const;
 
-const storage = BrowserStorage.create<Partial<GraphicsSettings>>('settings.graphics', {}, GraphicsSettingsSchema.partial());
+export const GraphicsSettingsStorage = BrowserStorage.create<Partial<GraphicsSettings>>('settings.graphics', {}, GraphicsSettingsSchema.partial());
 // Add a hook to purge the current graphics loader cache when the graphics settins change in any way
 // (most of the changes cause different textures to be loaded, so get rid of old ones)
-storage.subscribe(() => {
+GraphicsSettingsStorage.subscribe(() => {
 	// Do the cleanup asynchronously so things have time to unload if they were loaded
 	setTimeout(() => {
 		GraphicsManagerInstance.value?.loader.gc();
@@ -40,22 +38,22 @@ storage.subscribe(() => {
 });
 
 export function useGraphicsSettings(): GraphicsSettings {
-	const overrides = useObservable(storage);
+	const overrides = useObservable(GraphicsSettingsStorage);
 	return useMemo(() => ({
 		...GRAPHICS_SETTINGS_DEFAULT,
 		...overrides,
 	}), [overrides]);
 }
 
-function SetGraphicsSettings(changes: Partial<GraphicsSettings>): void {
-	storage.value = {
-		...storage.value,
+export function SetGraphicsSettings(changes: Partial<GraphicsSettings>): void {
+	GraphicsSettingsStorage.value = {
+		...GraphicsSettingsStorage.value,
 		...changes,
 	};
 }
 
-function ResetGraphicsSettings(settings: readonly (keyof GraphicsSettings)[]): void {
-	storage.produce((v) => {
+export function ResetGraphicsSettings(settings: readonly (keyof GraphicsSettings)[]): void {
+	GraphicsSettingsStorage.produce((v) => {
 		const newValue = CloneDeepMutable(v);
 		for (const s of settings) {
 			delete newValue[s];
@@ -64,34 +62,56 @@ function ResetGraphicsSettings(settings: readonly (keyof GraphicsSettings)[]): v
 	});
 }
 
-export function GraphicsSettings(): ReactElement | null {
-	return (
-		<>
-			<EffectsSettings />
-			<QualitySettings />
-			<GraphicsDebug />
-		</>
-	);
+//#region Setting-specific helpers and values
+
+// Smooth movement
+
+/**
+ * Tests whether graphics smooth movement should be enabled or not, if using "auto" option.
+ * @returns If smooth movement should be enabled by default
+ */
+export function useGraphicsSmoothMovementAutoEnabled(): boolean {
+	return !useMediaQuery('(prefers-reduced-motion)');
 }
 
-function EffectsSettings(): ReactElement {
-	const { effectBlinking } = useObservable(storage);
+/**
+ * Tests whether graphics smooth movement should be enabled or not, if using "auto" option.
+ * If it shouldn't be, it returns a short reason why not.
+ * @returns If smooth movement should be enabled by default, or reason why not
+ */
+export function useGraphicsSmoothMovementAutoEnabledExplain(): [true] | [false, reason: string] {
+	const prefersReducedMotion = useMediaQuery('(prefers-reduced-motion)');
 
-	return (
-		<fieldset>
-			<legend>Effects</legend>
-			<ToggleSettingInput
-				currentValue={ effectBlinking }
-				defaultValue={ GRAPHICS_SETTINGS_DEFAULT.effectBlinking }
-				label='Eye blinking of characters'
-				onChange={ (newValue) => {
-					SetGraphicsSettings({ effectBlinking: newValue });
-				} }
-				onReset={ () => ResetGraphicsSettings(['effectBlinking']) }
-			/>
-		</fieldset>
-	);
+	return useMemo((): ReturnType<typeof useGraphicsSmoothMovementAutoEnabledExplain> => {
+		if (prefersReducedMotion)
+			return [false, 'system settings prefer reduced motion'];
+
+		return [true];
+	}, [prefersReducedMotion]);
 }
+
+/**
+ * Checks whether graphics smooth movement should be enabled or not,
+ * checking user's settings and resolving "auto" value if necessary.
+ * @returns If smooth movement should be enabled
+ */
+export function useGraphicsSmoothMovementEnabled(): boolean {
+	// get storage value
+	const currentSmoothMovementSetting = useGraphicsSettings().smoothMovement;
+	const smoothMovementAutoSetting = useGraphicsSmoothMovementAutoEnabled();
+	// if auto call useGraphicsSmoothMovementAutoEnabled or useGraphicsSmoothMovementAutoEnabledExplain
+	switch (currentSmoothMovementSetting) {
+		case 'auto':
+			return smoothMovementAutoSetting;
+		case 'enabled':
+			return true;
+		case 'disabled':
+			return false;
+	}
+	AssertNever(currentSmoothMovementSetting);
+}
+
+// Texture resolution
 
 export const GRAPHICS_TEXTURE_RESOLUTION_SCALE: Record<Exclude<GraphicsSettings['textureResolution'], 'auto'>, number> = {
 	'1': 1,
@@ -99,153 +119,4 @@ export const GRAPHICS_TEXTURE_RESOLUTION_SCALE: Record<Exclude<GraphicsSettings[
 	'0.25': 4,
 };
 
-function QualitySettings(): ReactElement {
-	const { renderResolution, textureResolution, alphamaskEngine } = useObservable(storage);
-
-	const ALPHAMASK_ENGINES_DESCRIPTIONS: Record<GraphicsSettings['alphamaskEngine'], string> = {
-		pixi: 'Pixi.js',
-		customShader: 'Custom Pandora shader (recommended for less clipping)',
-		disabled: 'Ignore masks (recommended for avoiding lag)',
-	};
-
-	const automaticTextureResolution = useAutomaticResolution();
-
-	const GRAPHICS_TEXTURE_RESOLUTION_DESCRIPTIONS = useMemo((): Record<GraphicsSettings['textureResolution'], string | (() => string)> => ({
-		'auto': () => `Automatic (currently: ${ String(GRAPHICS_TEXTURE_RESOLUTION_DESCRIPTIONS[automaticTextureResolution]) })`,
-		'1': 'Full',
-		'0.5': '50%',
-		'0.25': '25%',
-	}), [automaticTextureResolution]);
-
-	return (
-		<fieldset>
-			<legend>Quality</legend>
-			<SelectSettingInput<string>
-				currentValue={ renderResolution?.toString() }
-				defaultValue={ GRAPHICS_SETTINGS_DEFAULT.renderResolution.toString() }
-				label='Render resolution'
-				stringify={
-					Object.fromEntries(
-						([100, 90, 80, 65, 50, 25, 0])
-							.map((v) => [v.toString(), `${v}%`]),
-					)
-				}
-				optionOrder={ [100, 90, 80, 65, 50, 25, 0].map(String) }
-				schema={ z.string() }
-				onChange={ (v) => {
-					const newValue = GraphicsSettingsSchema.shape.renderResolution.parse(Number.parseInt(v, 10));
-					SetGraphicsSettings({ renderResolution: newValue });
-				} }
-				onReset={ () => ResetGraphicsSettings(['renderResolution']) }
-			/>
-			<SelectSettingInput<GraphicsSettings['textureResolution']>
-				currentValue={ textureResolution }
-				defaultValue={ GRAPHICS_SETTINGS_DEFAULT.textureResolution }
-				label='Texture resolution'
-				stringify={ GRAPHICS_TEXTURE_RESOLUTION_DESCRIPTIONS }
-				optionOrder={ ['auto', '1', '0.5', '0.25'] }
-				schema={ GraphicsSettingsSchema.shape.textureResolution }
-				onChange={ (v) => SetGraphicsSettings({ textureResolution: v }) }
-				onReset={ () => ResetGraphicsSettings(['textureResolution']) }
-			/>
-			<SelectSettingInput<GraphicsSettings['alphamaskEngine']>
-				currentValue={ alphamaskEngine }
-				defaultValue={ GRAPHICS_SETTINGS_DEFAULT.alphamaskEngine }
-				label={
-					<>
-						Alphamasking engine
-						<ContextHelpButton>
-							<p>
-								Alphamasks allow assets to hide parts of other assets beneath them, so they do not clip through.<br />
-								Common examples of this are various shoes hiding feet/socks and pants hiding normally wide end of tops.<br />
-								While not using an alphamasking engine will cause clipping with assets, <br />
-								we decided to disable it by default, as many people are encountering performance issues due to it.
-							</p>
-
-							<p>
-								We have a planned rework that should allow alphamasks to work again without the performance impact, <br />
-								but it will take some time until this is implemented and adopted by assets.
-							</p>
-
-							<p>
-								If you do have a powerful computer, you can attempt to re-enable this option. <br />
-								In that case we recommend using the "Custom Pandora shader" option.
-							</p>
-						</ContextHelpButton>
-						<br />
-						<strong>The current implementation is known to cause lag on most devices</strong>
-					</>
-				}
-				stringify={ ALPHAMASK_ENGINES_DESCRIPTIONS }
-				schema={ GraphicsSettingsSchema.shape.alphamaskEngine }
-				onChange={ (v) => SetGraphicsSettings({ alphamaskEngine: v }) }
-				onReset={ () => ResetGraphicsSettings(['alphamaskEngine']) }
-			/>
-		</fieldset>
-	);
-}
-
-function GraphicsDebug(): ReactElement {
-	const graphicsManger = useObservable(GraphicsManagerInstance);
-
-	const [stats, setStats] = useState<IGraphicsLoaderStats>(() => ({
-		inUseTextures: 0,
-		loadedTextures: 0,
-		trackedTextures: 0,
-		loadedPixels: 0,
-		estLoadedSize: 0,
-	}));
-
-	const refreshStats = useCallback(() => {
-		if (graphicsManger == null)
-			return;
-
-		setStats(graphicsManger.loader.getDebugStats());
-	}, [graphicsManger]);
-
-	useEffect(() => {
-		if (graphicsManger == null)
-			return;
-
-		refreshStats();
-
-		return graphicsManger.loader.on('storeChaged', () => {
-			refreshStats();
-		});
-	}, [graphicsManger, refreshStats]);
-
-	if (graphicsManger == null) {
-		return (
-			<fieldset>
-				<legend>Graphics manager debug</legend>
-				Graphics manager is not loaded
-			</fieldset>
-		);
-	}
-
-	return (
-		<fieldset>
-			<legend>Graphics manager debug</legend>
-			<Column>
-				<Row>
-					<Button onClick={ refreshStats } slim>Refresh</Button>
-				</Row>
-				<span>Tracked textures: { stats.trackedTextures }</span>
-				<span>Loaded textures: { stats.loadedTextures }</span>
-				<span>Used textures: { stats.inUseTextures }</span>
-				<hr className='fill-x' />
-				<span>Loaded textures size (pixels): { stats.loadedPixels.toLocaleString(undefined, { useGrouping: 'always' }) }</span>
-				<span>Loaded textures size (bytes, estimate): { FormatBytes(stats.estLoadedSize, true) }</span>
-				<hr className='fill-x' />
-				<Row>
-					<Button onClick={ () => {
-						graphicsManger.loader.gc();
-						refreshStats();
-					} } slim>
-						Purge unused textures
-					</Button>
-				</Row>
-			</Column>
-		</fieldset>
-	);
-}
+//#endregion
